deleted one user class and updated
