+++ src/test/java/com/nextbasecrm/tests/US11_Aziza_CheckPopUpMessage.java
@@ -56,6 +56,31 @@ public class US11_Aziza_CheckPopUpMessage {
         Assert.assertEquals(actualPopupText,expectedPopUpText);
 
     }
+    @Test
+    public void errorMessage() {
+        //   String[] username ={"hr4@cydeo.com","hr5@cydeo.com","hr6@cydeo.com","helpdesk4@cydeo.com",
+        //  "helpdesk5@cydeo.com","helpdesk6@cydeo.com","marketing4@cydeo.com","marketing5@cydeo.com",
+        //  "marketing6@cydeo.com"};
+        //  for (String eachUserName : username) {
+        String username=ConfigurationReader.getProperty("username");
+        String password = ConfigurationReader.getProperty("password");
+        CRM_Utilities.crm_login(driver, username, password);
+        WebElement taskTab = driver.findElement(By.xpath("//div[@class ='feed-add-post-form-variants']//span[2]"));
+        taskTab.click();
+
+        WebElement iframe = driver.findElement(By.xpath("//div[@id='bx-html-editor-iframe-cnt-lifefeed_task_form']/iframe"));
+        driver.switchTo().frame(iframe);
+        BrowserUtils.sleep(3);
+        WebElement taskContentWindow = driver.findElement(By.xpath("//body[@contenteditable='true']"));
+        taskContentWindow.sendKeys("Aziza automated task 30");
+        driver.switchTo().defaultContent();
+        WebElement sendButton = driver.findElement(By.xpath("//button[@id='blog-submit-button-save']"));
+        sendButton.click();
+        String expectedErrorMessage = "The task name is not specified.";
+        String actualErrorMessage = driver.findElement(By.xpath("//div[@class='task-message-label error']")).getText();
+        Assert.assertTrue(expectedErrorMessage.equals(actualErrorMessage));
+
+    }
 
     @AfterMethod
     public void tearDown() {
